Add initial spaceship data to MongoDB
This commits adds a JavaScript file 'init-mongo.js', which creates a new sibling database 'spaceships', and a 'spaceship' collection in MongoDB. It then adds 10 entries to the database with fields such as name, franchise, and max speed of popular fiction spaceships. The file also includes checks to prevent re-initialization of the database.