Refactor Spaceships controller tests and simplify update method
The tests for the SpaceshipsGetController have been refactored to replace the now-deprecated SpaceshipsGetter with the new SpaceshipsByCriteriaSearcher. This introduces a new testing pattern following the change from PageResponse objects to ListResponse objects. The update method in the SpaceshipsPutController has been simplified to not mock saving of the spaceship object, as it wasn't required for the test.